remove unused MC_Scanner::loc
It looks as if this variable is used, but actually it is not. The #define to
YY_DECL differs from the MC_Scanner::yylex prototype and, as a result, the
second argument to this method shadows the class member. It does not seem as if
any code relies on MC_Scanner::loc existing, so this commit removes it for
clarity.